+++ src/main/java/com/influxdb/v3/client/internal/InfluxDBClientImpl.java
@@ -38,6 +38,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import io.netty.handler.codec.http.HttpMethod;
+import org.apache.arrow.flight.CallOption;
 import org.apache.arrow.vector.FieldVector;
 import org.apache.arrow.vector.VectorSchemaRoot;
 
@@ -341,7 +342,15 @@ public final class InfluxDBClientImpl implements InfluxDBClient {
             }
         });
 
-        return flightSqlClient.execute(query, database, options.queryTypeSafe(), parameters, options.headersSafe());
+        CallOption[] callOptions = options.grpcCallOptions().getCallOptions();
+        return flightSqlClient.execute(
+                query,
+                database,
+                options.queryTypeSafe(),
+                parameters,
+                options.headersSafe(),
+                callOptions
+        );
     }
 
     @Nonnull